Move Objects -> Support, remove private constructor
