+++ tests/Support/DefinitionStorage/ServiceWithBuiltinTypeWithoutDefault.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Definitions\Tests\Objects\DefinitionStorage;
+namespace Yiisoft\Definitions\Tests\Support\DefinitionStorage;
 
 final class ServiceWithBuiltinTypeWithoutDefault
 {
